refactor: begin moving proptypes.js files to prop-types.js
