+++ packages/article/src/article-lead-asset/crop-prop-types.js
